dev-requirements.txt
-e .
flake8<3.0
invoke>=0.11.1,<0.12
pytest
pytest-cov
pytest-django<3.0.0
sphinx
sphinx_rtd_theme
tox
twine
wheel
